# openapi.yaml
openapi: 3.0.3
info:
  title: Pafodex Backend API
  version: 1.0.0
  description: |
    Documentation Swagger de l'API Pafodex.
    Les endpoints exposés sont détaillés avec les payloads, paramètres, réponses et pagination.
servers:
  - url: /
    description: Serveur courant
  - url: https://pafodex-backend.onrender.com
    description: Serveur Render
paths:
  /api/register:
    post:
      tags:
        - Auth
      summary: Enregistrer un nouvel utilisateur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Utilisateur créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Requête invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email déjà utilisé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/login:
    post:
      tags:
        - Auth
      summary: Authentifier un utilisateur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentification réussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Requête invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Identifiants invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/logout:
    post:
      tags:
        - Auth
      summary: Déconnecter l'utilisateur
      description: |
        Le token peut être fourni dans l'en-tête Authorization: Bearer <token>
        ou en query paramètre apiToken.
      parameters:
        - $ref: '#/components/parameters/ApiTokenQuery'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Déconnexion réussie
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '401':
          description: Token manquant ou invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/me:
    get:
      tags:
        - Auth
      summary: Récupérer les informations de l'utilisateur connecté
      description: |
        Le token peut être fourni dans l'en-tête Authorization: Bearer <token>
        ou en query paramètre apiToken.
      parameters:
        - $ref: '#/components/parameters/ApiTokenQuery'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Informations utilisateur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          description: Token manquant ou invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Auth
      summary: Mettre à jour l'utilisateur connecté
      description: |
        Le token peut être fourni dans l'en-tête Authorization: Bearer <token>
        ou en query paramètre apiToken.
      parameters:
        - $ref: '#/components/parameters/ApiTokenQuery'
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMeRequest'
      responses:
        '200':
          description: Utilisateur mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '400':
          description: Requête invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Token manquant ou invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email déjà utilisé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Auth
      summary: Supprimer l'utilisateur connecté
      description: |
        Le token peut être fourni dans l'en-tête Authorization: Bearer <token>
        ou en query paramètre apiToken.
      parameters:
        - $ref: '#/components/parameters/ApiTokenQuery'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Utilisateur supprimé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Token manquant ou invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Bearer
  parameters:
    ApiTokenQuery:
      name: apiToken
      in: query
      description: Token d'API alternatif lorsque Authorization n'est pas fourni
      required: false
      schema:
        type: string
  schemas:
    RegisterRequest:
      type: object
      required:
        - pseudo
        - mail
        - password
        - passwordConfirm
      properties:
        pseudo:
          type: string
          example: user123
        mail:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          example: password123
        passwordConfirm:
          type: string
          format: password
          example: password123
    RegisterResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        apiToken:
          type: string
          example: abcdef0123456789abcdef0123456789abcdef0123456789abcdef0123456789
    LoginRequest:
      type: object
      required:
        - mail
        - password
      properties:
        mail:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          example: password123
    LoginResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        apiToken:
          type: string
          example: abcdef0123456789abcdef0123456789abcdef0123456789abcdef0123456789
    MeResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        pseudo:
          type: string
          example: user123
        mail:
          type: string
          format: email
          example: user@example.com
    UpdateMeRequest:
      type: object
      properties:
        pseudo:
          type: string
          example: user123
        mail:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          example: newPassword123
        passwordConfirm:
          type: string
          format: password
          example: newPassword123
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: Operation completed successfully
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Missing fields